requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.41.2
flasgger==0.9.7.1
pytest==8.3.3